requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dash==3.2.0
dparse==0.6.4
et_xmlfile==2.0.0
fastapi==0.120.0
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
matplotlib==3.10.7
mdurl==0.1.2
narwhals==2.9.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.3.1
psutil==7.1.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.10.23
requests==2.32.5
retrying==1.4.2
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safety==3.6.2
safety-schemas==0.0.16
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
